Simplify the logic by using a closure to store the extension state.
Before, we used an object `{ body, extension }` to store the value
+ extension. Part of the reason for this was so that we would not
have to create a new closure for the proxy handlers object every
time we create a proxy. However, due to the invariants of Proxy
this led to having to define a bunch more traps, and handle a lot
of edge cases. The complexity wasn't worth it, so changing it to
just use a closure (which is faster in the long run anyway).